CLN: move BaseCVSWriter* classes to notification_listeners module
